// src/app/shared/model/pages.model.ts
export interface levelUserFormField {
  userLevelId: string;
  formFieldId: {
    errorMessages: {
      required: string;
      minLength?: string;
      maxLength?: string;
      regex?: string;
    };
    name: string;
    label: string;
    inputType: string;
    placeholder?: string;
    minLength?: number;
    maxLength?: number;
    regex?: string;
    required?: boolean;
    status?: string;
    uuid: string;
    options?: { label: string; value: any }[];
  };
  isRequired: boolean;
  isEnabled: boolean;
  order: number;
  uuid: string;
}